Revert "test BF16 expert affinities"
This reverts commit 5a287553f51de7bf415c2a136999796e115a7b2d.

+++ vLLM-neuron/__init__.py
@@ -62,7 +62,7 @@ def _patch_qwen3_moe(module):
             nki_affinities != 0,
             router_probs,
             torch.zeros_like(router_probs),
-        ).to(torch.bfloat16)
+        )
         if self.world_size > 1:
             expert_affinities = self.tp_group.all_gather(
                 expert_affinities, dim=0
